requirements.txt
cffi==2.0.0
cryptography==46.0.4
fernet==1.0.1
mysql-connector-python==9.5.0
pyaes==1.6.1
pycparser==3.0
